% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate.R
\name{aggregate_emmissions}
\alias{aggregate_emmissions}
\title{Aggregates emission budget data}
\usage{
aggregate_emmissions(
  data,
  subsector_new = NULL,
  cols = c(Dienstleistungen = "Gold", Haushalte = "Green", Industrie = "Blue",
    `Land- und Forstw.` = "Purple", Verkehr = "Gray", `natürl. Emissionen` = "natural")
)
}
\arguments{
\item{cols}{}
}
\description{
Merges subsectors with small emissions into one common (or pre-defined) category and sums up emissions accordingly.
}
