% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate.R
\name{aggregate_map}
\alias{aggregate_map}
\title{Aggregates sf multipolygon into one bounding polygon}
\usage{
aggregate_map(map)
}
\arguments{
\item{map}{}
}
\description{
Aggregates sf multipolygon into one bounding polygon
}
