% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate.R
\name{aggregate_ndep_exposition_distrib}
\alias{aggregate_ndep_exposition_distrib}
\title{Aggregates sensitive ecosystem exposition towards reactive atmospheric nitrogen deposition dataset}
\usage{
aggregate_ndep_exposition_distrib(data)
}
\arguments{
\item{data}{}
}
\description{
Calculates number of sensitive ecosystems per bin of maximum nitrogen deposition exceeding ecosystem-specific Critical Loads and year;
number of sensitive ecosystems per bin, cumulative number of sensitive ecosystems per bin and relative cumulative number of sensitive ecosystems per bin.
}
