% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate.R
\name{aggregate_population_exposition_distrib}
\alias{aggregate_population_exposition_distrib}
\title{Aggregates air pollutant population exposition dataset}
\usage{
aggregate_population_exposition_distrib(data)
}
\arguments{
\item{data}{}
}
\description{
Calculates sum of inhabitant population per pollutant concentration bins and years;
derives population per bin, cumulative population per bin and relative cumulative population per bin.
}
