% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate.R
\name{aggregate_population_weighted_mean}
\alias{aggregate_population_weighted_mean}
\title{Aggregates to mean population-weighted concentrations from air pollutant and inhabitant population dataset}
\usage{
aggregate_population_weighted_mean(
  data,
  groups = c("year", "pollutant", "metric", "parameter")
)
}
\arguments{
\item{data}{}

\item{groups}{}
}
\description{
Aggregates to mean population-weighted concentrations from air pollutant and inhabitant population dataset
}
