% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_helpers.R
\name{aggregate_rsd}
\alias{aggregate_rsd}
\title{Aggregate emissions from RSD dataset using some statistical functions}
\usage{
aggregate_rsd(
  data,
  meta,
  y = "nox_emission",
  groups = c("vehicle_type", "vehicle_fuel_type", "vehicle_euronorm"),
  nmin = 50
)
}
\arguments{
\item{data}{}

\item{meta}{}

\item{y}{}

\item{groups}{}

\item{nmin}{}
}
\description{
Aggregate emissions from RSD dataset using some statistical functions
}
\keyword{internal}
