% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate.R
\name{aggregate_rsd_nox}
\alias{aggregate_rsd_nox}
\title{Aggregates RSD NOx emissions to specified groups}
\usage{
aggregate_rsd_nox(
  data,
  rsd_auxiliary,
  groups = c("vehicle_type", "vehicle_fuel_type", "vehicle_euronorm")
)
}
\arguments{
\item{groups}{}
}
\description{
Combines measurement data with auxiliary metadata, filters dataset using provided filter criteria
and calculates mean values of specified vehicle groups.
}
