% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/airquality.methods-package.R
\docType{package}
\name{airquality.methods-package}
\alias{airquality.methods}
\alias{airquality.methods-package}
\title{airquality.methods: Methods for analysing air quality in the Canton of Zurich}
\description{
Functions for analysis R-project "airquality" concerning a systematic and comprehensive air quality data compilation, analysis and visualisation for the Canton of Zurich, based on public data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/awelZH/airquality.methods}
  \item Report bugs at \url{https://github.com/awelZH/airquality.methods/issues}
}

}
\author{
\strong{Maintainer}: Jörg Sintermann \email{luft@bd.zh.ch} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

Authors:
\itemize{
  \item Jörg Sintermann \email{luft@bd.zh.ch} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})
}

}
\keyword{internal}
