% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo-admin-stac.R
\name{asset_compression}
\alias{asset_compression}
\title{Derive the compression from an asset url}
\usage{
asset_compression(href)
}
\arguments{
\item{href}{Asset url.}
}
\value{
\code{"zip"}, \code{"gz"} or \code{NA}.
}
\description{
Derive the compression from an asset url
}
\keyword{internal}
