% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo-admin-stac.R
\name{asset_format}
\alias{asset_format}
\title{Derive the file format from an asset url}
\usage{
asset_format(href)
}
\arguments{
\item{href}{Asset url.}
}
\value{
Character vector of formats.
}
\description{
\code{".csv.zip"} becomes \code{"csv"}, \code{".tiff"} becomes \code{"tif"}.
}
\keyword{internal}
