% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_helpers.R
\name{average_to_grid}
\alias{average_to_grid}
\title{Average stars raster data to another stars grid as mean values}
\usage{
average_to_grid(data, grid, method = "average", na_val = -999)
}
\arguments{
\item{na_val}{}
}
\description{
Average stars raster data to another stars grid as mean values
}
\keyword{internal}
