% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_helpers.R
\name{average_to_statpop}
\alias{average_to_statpop}
\title{Average stars raster data to the grid of BFS statpop dataset}
\usage{
average_to_statpop(x, y)
}
\arguments{
\item{y}{}
}
\description{
Average stars raster data to the grid of BFS statpop dataset
}
\keyword{internal}
