% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.R
\name{bafu_rasterlist_to_tibble}
\alias{bafu_rasterlist_to_tibble}
\title{Converts list of stars raster data to tibble data (for BAFU raster data)}
\usage{
bafu_rasterlist_to_tibble(rasterlist)
}
\arguments{
\item{rasterlist}{}
}
\description{
Converts list of stars raster data to tibble data (for BAFU raster data)
}
\keyword{internal}
