% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo-admin-raster.R
\name{bbox_ch_lv95}
\alias{bbox_ch_lv95}
\title{Default extent: Switzerland in LV95}
\format{
Named numeric vector of length 4 (\code{xmin}, \code{ymin}, \code{xmax}, \code{ymax}).
}
\usage{
bbox_ch_lv95
}
\description{
Rounded to 10 km with a margin.
}
