% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_helpers.R
\name{bin_concentration}
\alias{bin_concentration}
\title{Calculate concentration bins for exposition assessment}
\usage{
bin_concentration(data)
}
\arguments{
\item{data}{}
}
\description{
Calculate concentration bins for exposition assessment
}
\keyword{internal}
