% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_helpers.R
\name{calc_population_weighted_mean}
\alias{calc_population_weighted_mean}
\title{Calculate population weighted mean concentration for exposition assessment}
\usage{
calc_population_weighted_mean(concentration, population)
}
\arguments{
\item{concentration}{}

\item{population}{}
}
\description{
Calculate population weighted mean concentration for exposition assessment
}
\keyword{internal}
