% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_helpers.R
\name{calc_rsd_nox_emission}
\alias{calc_rsd_nox_emission}
\title{Calculate RSD NOx emissions in g/kg fuel}
\usage{
calc_rsd_nox_emission(NO, p, CO2, CO, HC)
}
\arguments{
\item{NO}{}

\item{p}{}

\item{CO2}{}

\item{CO}{}

\item{HC}{}
}
\description{
Calculate RSD NOx emissions in g/kg fuel
}
\keyword{internal}
