% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_helpers.R
\name{calc_vsp}
\alias{calc_vsp}
\title{Calculate vehicle specific power following Jiménez}
\usage{
calc_vsp(
  speed,
  accel,
  slope,
  vsp.a = 1.1,
  vsp.b = 0.132,
  vsp.c = 0.000302,
  vsp.g = 9.81
)
}
\arguments{
\item{vsp.g}{}
}
\description{
Calculate vehicle specific power following Jiménez
}
\keyword{internal}
