% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster-align.R
\name{check_year_match}
\alias{check_year_match}
\title{Validate the year_match argument}
\usage{
check_year_match(year_match, call = rlang::caller_env())
}
\arguments{
\item{year_match}{One value or a named vector of values.}

\item{call}{Environment used for the error call.}
}
\value{
\code{TRUE}, invisibly.
}
\description{
Validate the year_match argument
}
\keyword{internal}
