% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ndep.R
\name{classify_frac_estimated}
\alias{classify_frac_estimated}
\title{Class of the estimated fraction of the nitrogen deposition}
\usage{
classify_frac_estimated(
  frac_estimated,
  breaks = c(0, 0.33, 0.66, 1),
  labels = c("<33\%", "33-66\%", ">66\%")
)
}
\arguments{
\item{frac_estimated}{Estimated part divided by the total deposition (0-1), not the estimated
deposition itself.}

\item{breaks, labels}{Class boundaries and labels; the lower bound of the first class is included.}
}
\value{
Factor "<33\%", "33-66\%", ">66\%".
}
\description{
Class of the estimated fraction of the nitrogen deposition
}
\examples{
classify_frac_estimated(c(0.1, 0.5, 0.9))

}
