% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ndep.R
\name{classify_nh3_emission}
\alias{classify_nh3_emission}
\title{NH3 emission class of a site (emissions within 1 km)}
\usage{
classify_nh3_emission(
  emission,
  breaks = c(0, 10, 30, Inf),
  labels = c("tief", "mittel", "hoch")
)
}
\arguments{
\item{emission}{NH3 emission within 1 km.}

\item{breaks, labels}{Class boundaries and labels; the lower bound of the first class is included.}
}
\value{
Factor "tief", "mittel", "hoch"; \code{NA} where \code{emission} is \code{NA}.
}
\description{
NH3 emission class of a site (emissions within 1 km)
}
\examples{
classify_nh3_emission(c(5, 20, 40))

}
