% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo-admin-stac.R
\name{collection_label}
\alias{collection_label}
\title{Short label for a collection id}
\usage{
collection_label(collection)
}
\arguments{
\item{collection}{Collection id.}
}
\value{
A character vector of labels.
}
\description{
\code{"ch.bafu.luftreinhaltung-x_y"} becomes \code{"luftreinhaltung_x_y"}.
}
\keyword{internal}
