% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster-align.R
\name{combine_layers}
\alias{combine_layers}
\title{Assemble attribute arrays into a stars object on the template}
\usage{
combine_layers(layers, template, call = rlang::caller_env())
}
\arguments{
\item{layers}{Named list of arrays.}

\item{template}{The \code{stars} template.}

\item{call}{Environment used for the error call.}
}
\value{
A \code{stars} object.
}
\description{
Assemble attribute arrays into a stars object on the template
}
\keyword{internal}
