% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_helpers.R
\name{combine_raster_aq}
\alias{combine_raster_aq}
\title{just to bring ist into required structure}
\usage{
combine_raster_aq(
  years,
  data_raster_pm25,
  data_raster_pm10,
  data_raster_no2,
  data_raster_o3mp98,
  data_raster_ndep
)
}
\arguments{
\item{data_raster_ndep}{}
}
\description{
just to bring ist into required structure
}
\keyword{internal}
