% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ndep.R
\name{derive_source_category}
\alias{derive_source_category}
\title{Source category of a nitrogen deposition parameter}
\usage{
derive_source_category(parameter)
}
\arguments{
\item{parameter}{Parameter names, e.g. "NO2-N_ddep" or "NH4-N_bdep".}
}
\value{
"aus NOx-Quellen" (names containing "NO"), "aus NH3-Quellen" (names containing "NH")
or \code{NA}.
}
\description{
Source category of a nitrogen deposition parameter
}
\examples{
derive_source_category(c("NO2-N_ddep", "NH4-N_bdep"))

}
