% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{derive_trends_per_parameter}
\alias{derive_trends_per_parameter}
\title{Wrapper to derive site-specific statistical meteo-normalisation trends per parameter from pre-compiled monitoring data including aggregation to yearly trend values}
\usage{
derive_trends_per_parameter(
  data_trends,
  parameter,
  reference_year_fun,
  yearmin_per_site = 4,
  frac_train = 0.8,
  ntrees = 300,
  nsamples = 300,
  verbose = TRUE,
  minimal = TRUE,
  coverage = 0.8
)
}
\arguments{
\item{coverage}{}
}
\description{
Wrapper to derive site-specific statistical meteo-normalisation trends per parameter from pre-compiled monitoring data including aggregation to yearly trend values
}
