% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basemap-swisstopo.R
\name{.basemap_check_png}
\alias{.basemap_check_png}
\title{Verify that a file really is a PNG}
\usage{
.basemap_check_png(path, layer, width, height)
}
\description{
Verify that a file really is a PNG
}
\keyword{internal}
