% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basemap-swisstopo.R
\name{.basemap_read_worldfile}
\alias{.basemap_read_worldfile}
\title{Read a world file back into a bbox}
\usage{
.basemap_read_worldfile(path, width, height)
}
\description{
Read a world file back into a bbox
}
\keyword{internal}
