% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basemap-swisstopo.R
\name{.basemap_worldfile}
\alias{.basemap_worldfile}
\title{World file contents for a bbox}
\usage{
.basemap_worldfile(bbox, width, height)
}
\description{
World file contents for a bbox
}
\keyword{internal}
