% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale-capped-limits.R
\name{.capped_check_probs}
\alias{.capped_check_probs}
\title{Check and normalise \code{probs}}
\usage{
.capped_check_probs(probs)
}
\description{
Check and normalise \code{probs}
}
\keyword{internal}
