% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale-capped-guide.R
\name{.capped_guide_params}
\alias{.capped_guide_params}
\title{Extra parameters the capped guides understand}
\usage{
.capped_guide_params
}
\description{
Extra parameters the capped guides understand
}
\keyword{internal}
