% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale-capped-guide.R
\name{.capped_key_add}
\alias{.capped_key_add}
\title{Append a row to a guide key while preserving attributes}
\usage{
.capped_key_add(key, value, label, colour, aes_col)
}
\description{
Append a row to a guide key while preserving attributes
}
\keyword{internal}
