% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polar-raster-grid.R
\name{.polar_compass}
\alias{.polar_compass}
\title{Resolve compass directions into a named angle vector}
\usage{
.polar_compass(compass)
}
\description{
Resolve compass directions into a named angle vector
}
\keyword{internal}
