% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polar-raster-prepare.R
\name{.polar_data}
\alias{.polar_data}
\title{Get the data out of an openair object or a data frame}
\usage{
.polar_data(x)
}
\description{
Get the data out of an openair object or a data frame
}
\keyword{internal}
