% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polar-raster-axis.R
\name{.polar_el}
\alias{.polar_el}
\title{Look up a theme element, tolerating incomplete themes}
\usage{
.polar_el(theme, name)
}
\description{
Look up a theme element, tolerating incomplete themes
}
\keyword{internal}
