% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polar-map-sites.R
\name{.polar_map_closest}
\alias{.polar_map_closest}
\title{Closest pair of sites, as names and distance}
\usage{
.polar_map_closest(centres)
}
\description{
Closest pair of sites, as names and distance
}
\keyword{internal}
