% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polar-map-decor.R
\name{.polar_map_corner}
\alias{.polar_map_corner}
\title{Corner position for a piece of furniture}
\usage{
.polar_map_corner(pos, xlim, ylim, half_w, half_h, pad)
}
\description{
Corner position for a piece of furniture
}
\keyword{internal}
