% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polar-map-decor.R
\name{.polar_map_disc}
\alias{.polar_map_disc}
\title{A filled disc, used as the plaque behind the key}
\usage{
.polar_map_disc(cx, cy, r)
}
\description{
A filled disc, used as the plaque behind the key
}
\keyword{internal}
