% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polar-map-decor.R
\name{.polar_map_sites_layers}
\alias{.polar_map_sites_layers}
\title{Site markers and name labels}
\usage{
.polar_map_sites_layers(decor, centres, radius, expand)
}
\description{
Site markers and name labels
}
\keyword{internal}
