% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_helpers.R
\name{download_statpop_data}
\alias{download_statpop_data}
\title{Download statpop *.zip from BFS api}
\usage{
download_statpop_data(year, destination_path, file_filter = NULL)
}
\arguments{
\item{file_filter}{}
}
\description{
Download statpop *.zip from BFS api
}
\keyword{internal}
