% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/municipalities.R
\name{drop_foreign_enclaves}
\alias{drop_foreign_enclaves}
\title{Remove enclaves of other cantons from the municipality map}
\usage{
drop_foreign_enclaves(map_municipalities)
}
\arguments{
\item{map_municipalities}{\code{sf} polygons with \code{art_text}.}
}
\value{
\code{map_municipalities} without enclaves of other cantons.
}
\description{
The geolion map contains the Kloster Fahr, an enclave of the Canton of Aargau
(\code{bfs = 0}, "ausserkantonale Enklave"). It is not part of the canton.
}
