% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{expositionpars}
\alias{expositionpars}
\title{Wrapper to supply pollutant-specific list of parameters for exposition plotting}
\usage{
expositionpars(parameter)
}
\arguments{
\item{parameter}{}
}
\description{
Wrapper to supply pollutant-specific list of parameters for exposition plotting
}
\keyword{internal}
