% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{extract_pollutant}
\alias{extract_pollutant}
\title{Extract polutant from BAFU data.geo.admin.ch ressource string}
\usage{
extract_pollutant(id)
}
\arguments{
\item{id}{}
}
\description{
Extract polutant from BAFU data.geo.admin.ch ressource string
}
\keyword{internal}
