% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_helpers.R
\name{filter_rsd}
\alias{filter_rsd}
\title{Filter RSD dataset using filter criteria}
\usage{
filter_rsd(data, filters)
}
\arguments{
\item{filters}{}
}
\description{
Filter RSD dataset using filter criteria
}
\keyword{internal}
