% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo-admin-cache.R
\name{geo_admin_cache_dir}
\alias{geo_admin_cache_dir}
\title{Default cache directory}
\usage{
geo_admin_cache_dir()
}
\value{
Path to the cache directory.
}
\description{
Persists between R sessions, so repeated analyses run offline.
}
