% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_helpers.R
\name{get_assets}
\alias{get_assets}
\title{read wcs layer from geolion}
\usage{
get_assets(metadata, filter)
}
\arguments{
\item{na_value}{}

\item{cov_stack}{}
}
\description{
read wcs layer from geolion
}
\keyword{Convert}
\keyword{a}
\keyword{coverage}
\keyword{data.frame}
\keyword{geolion}
\keyword{internal}
\keyword{list}
\keyword{to}
\keyword{wcs}
