% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_helpers.R
\name{get_bfs_metadata}
\alias{get_bfs_metadata}
\title{Get dataset download url from BFS api based on BFS dataset id}
\usage{
get_bfs_metadata(bfs_nr)
}
\arguments{
\item{bfs_nr}{}
}
\description{
Get dataset download url from BFS api based on BFS dataset id
}
\keyword{internal}
