% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-tabular.R
\name{get_opendataswiss_metadata}
\alias{get_opendataswiss_metadata}
\title{Get the download links of an opendata.swiss dataset}
\usage{
get_opendataswiss_metadata(
  apiurl,
  file_filter = ".csv",
  useragent = "Amt für Abfall, Wasser, Energie und Luft, Kanton Zürich"
)
}
\arguments{
\item{apiurl}{Package-show url of the CKAN API.}

\item{file_filter}{Substring the download url must contain.}

\item{useragent}{User agent sent with the request.}
}
\value{
Character vector of download urls.
}
\description{
Get the download links of an opendata.swiss dataset
}
\keyword{internal}
