% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_years}
\alias{get_years}
\title{Check previously analysed years and determine which new years should be analysed}
\usage{
get_years(read_all_raster, yearmax, base_scenario_year)
}
\arguments{
\item{yearmax}{}
}
\description{
Check previously analysed years and determine which new years should be analysed
}
\keyword{internal}
