% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{ggplot_timeseries}
\alias{ggplot_timeseries}
\title{Plot timeseries yearly monitoring-data using ggplot2}
\usage{
ggplot_timeseries(
  data,
  mapping = ggplot2::aes(x = year, y = concentration, color = siteclass),
  ylims = c(0, NA),
  ybreaks = waiver(),
  titlelab = NULL,
  captionlab = NULL,
  pointshape = 19,
  pointsize = 2,
  threshold = list(value = NA, color = "gray30", label = NULL, labelsize = 4, linetype =
    2, linesize = 1),
  theme = ggplot2::theme_minimal()
)
}
\arguments{
\item{data}{}

\item{mapping}{}

\item{ylims}{}

\item{ybreaks}{}

\item{titlelab}{}

\item{captionlab}{}

\item{pointshape}{}

\item{pointsize}{}

\item{threshold}{}

\item{theme}{}
}
\description{
Plot timeseries yearly monitoring-data using ggplot2
}
\keyword{internal}
