% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{ggplot_timeseries_lines}
\alias{ggplot_timeseries_lines}
\title{Plot timeseries using lines}
\usage{
ggplot_timeseries_lines(
  data,
  mapping = ggplot2::aes(x = year, y = population_weighted_mean),
  ylims = c(0, NA),
  ybreaks = waiver(),
  titlelab = NULL,
  captionlab = NULL,
  theme = ggplot2::theme_minimal()
)
}
\arguments{
\item{theme}{}
}
\description{
Plot timeseries using lines
}
\keyword{internal}
