% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{immissionscale_parameters}
\alias{immissionscale_parameters}
\title{Parameters that \code{\link[=immissionscale]{immissionscale()}} knows}
\usage{
immissionscale_parameters()
}
\value{
Character vector of parameter codes.
}
\description{
Parameters that \code{\link[=immissionscale]{immissionscale()}} knows
}
\examples{
immissionscale_parameters()

}
