% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode.R
\name{longparameter}
\alias{longparameter}
\title{Short metric label for a parameter code}
\usage{
longparameter(x)
}
\arguments{
\item{x}{Character vector of parameter codes.}
}
\value{
Character vector of short labels.
}
\description{
The compact variant used inside plot annotations, where space is tight.
Anything that is not an explicit ozone metric is an annual mean.
}
\examples{
longparameter(c("O3_max_98p_m1", "NO2"))

}
