% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_helpers.R
\name{merge_restructure_rsd}
\alias{merge_restructure_rsd}
\title{Merge RSD data with corresponding metadata}
\usage{
merge_restructure_rsd(data, meta)
}
\arguments{
\item{meta}{}
}
\description{
Merge RSD data with corresponding metadata
}
\keyword{internal}
