% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_helpers.R
\name{merge_statpop_with_municipalities}
\alias{merge_statpop_with_municipalities}
\title{Merge air quality and statpop rasterdata with municipilty boundaries and convert to a common tibble}
\usage{
merge_statpop_with_municipalities(data_raster, data_municip)
}
\arguments{
\item{data_municip}{}
}
\description{
Merge air quality and statpop rasterdata with municipilty boundaries and convert to a common tibble
}
\keyword{internal}
