% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_helpers.R
\name{merge_statpop_with_subareas}
\alias{merge_statpop_with_subareas}
\title{Merge air quality and statpop rasterdata with municipilty boundaries and convert to a common tibble}
\usage{
merge_statpop_with_subareas(
  data_raster,
  data_subareas,
  join_by = "bfs",
  id_subareas = "gemeindename"
)
}
\arguments{
\item{id_subareas}{}
}
\description{
Merge air quality and statpop rasterdata with municipilty boundaries and convert to a common tibble
}
\keyword{internal}
