% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo-admin-stac.R
\name{next_stac_page}
\alias{next_stac_page}
\title{Follow the STAC link with \code{rel = "next"}}
\usage{
next_stac_page(resp, req)
}
\arguments{
\item{resp}{The last response.}

\item{req}{The original request.}
}
\value{
A request for the next page, or \code{NULL} when the last page is reached.
}
\description{
Interface expected by \code{\link[httr2:req_perform_iterative]{httr2::req_perform_iterative()}}.
}
\keyword{internal}
