% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_all_expo_cumul}
\alias{plot_all_expo_cumul}
\title{Wrapper to plot pollutant-specific cumulated population exposition distribution in a nested list by pollutant & year}
\usage{
plot_all_expo_cumul(parameter, data)
}
\arguments{
\item{parameter}{}

\item{data}{}
}
\description{
Wrapper to plot pollutant-specific cumulated population exposition distribution in a nested list by pollutant & year
}
\keyword{internal}
