% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_all_expo_hist}
\alias{plot_all_expo_hist}
\title{Wrapper to plot pollutant-specific population exposition histograms in a nested list by pollutant & year}
\usage{
plot_all_expo_hist(parameter, data, sub = "im Kanton Zürich")
}
\arguments{
\item{data}{}
}
\description{
Wrapper to plot pollutant-specific population exposition histograms in a nested list by pollutant & year
}
\keyword{internal}
