% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_all_expo_hist_ndep}
\alias{plot_all_expo_hist_ndep}
\title{Wrapper to plot sensitive ecosystem nitrogen deposition exposition histograms in a list by year}
\usage{
plot_all_expo_hist_ndep(data, threshold_ndep)
}
\arguments{
\item{data}{}

\item{threshold_ndep}{}
}
\description{
Wrapper to plot sensitive ecosystem nitrogen deposition exposition histograms in a list by year
}
\keyword{internal}
