% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_all_popweighmean_maps}
\alias{plot_all_popweighmean_maps}
\title{Wrapper to plot pollutant-specific population weighted mean maps by municipality in a nested list by pollutant & year}
\usage{
plot_all_popweighmean_maps(parameter, data, data_canton)
}
\arguments{
\item{data_canton}{}
}
\description{
Wrapper to plot pollutant-specific population weighted mean maps by municipality in a nested list by pollutant & year
}
\keyword{internal}
