% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_pars_monitoring_timeseries}
\alias{plot_pars_monitoring_timeseries}
\title{Wrapper to plot timeseries yearly data using ggplot2 providing pollutant-specific list}
\usage{
plot_pars_monitoring_timeseries(data, parameters)
}
\arguments{
\item{data}{}

\item{parameters}{}
}
\description{
Wrapper to plot timeseries yearly data using ggplot2 providing pollutant-specific list
}
\keyword{internal}
