% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polar-raster-plot.R
\name{polar_data}
\alias{polar_data}
\title{Get the grid data out of a plot made by polar_plot()}
\usage{
polar_data(p)
}
\arguments{
\item{p}{Result of \code{\link[=polar_plot]{polar_plot()}}.}
}
\value{
The \code{u}/\code{v}/\code{z} data frame, or \code{NULL}.
}
\description{
Get the grid data out of a plot made by polar_plot()
}
