% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.R
\name{prepare_emission_trends}
\alias{prepare_emission_trends}
\title{Wrangle air pollutant emission data into relative emission trends for according comparison with monitoring data and trends}
\usage{
prepare_emission_trends(emissions, reference_year_fun)
}
\arguments{
\item{reference_year_fun}{}
}
\description{
Wrangle air pollutant emission data into relative emission trends for according comparison with monitoring data and trends
}
