% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.R
\name{prepare_emmissions}
\alias{prepare_emmissions}
\title{Prepares emission dataset for appropriate aggregation}
\usage{
prepare_emmissions(
  data,
  filter_args = stand == max(stand) & canton == "ZH" & emission != 0 & !(subsector \%in\%
    c("Weitere Punktquellen OL", "Rheinschifffahrt", "Flugverkehr Genf"))
)
}
\arguments{
\item{filter_args}{}
}
\description{
Prepares emission dataset for appropriate aggregation
}
