% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.R
\name{prepare_exposition}
\alias{prepare_exposition}
\title{Convert BFS statpop spatial raster rasterdata as well as air quality spatial rasterdata to combined tibble dataset for use in script}
\usage{
prepare_exposition(data_raster_bfs, data_raster_aq, years)
}
\arguments{
\item{years}{}
}
\description{
Convert BFS statpop spatial raster rasterdata as well as air quality spatial rasterdata to combined tibble dataset for use in script
}
