% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.R
\name{prepare_preliminary_deaths}
\alias{prepare_preliminary_deaths}
\title{Prepare preliminary deaths from population-weighted mean, mortality cases and outcome metadata and derive health outcomes}
\usage{
prepare_preliminary_deaths(
  data_expo_weighmean,
  data_mortality,
  outcomes_meta,
  conc_threshold = "lower_conc_threshold"
)
}
\arguments{
\item{data_expo_weighmean}{}

\item{data_mortality}{}

\item{outcomes_meta}{}

\item{conc_threshold}{}
}
\description{
Prepare preliminary deaths from population-weighted mean, mortality cases and outcome metadata and derive health outcomes
}
