% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.R
\name{prepare_rasterdata_aq_base}
\alias{prepare_rasterdata_aq_base}
\title{Spatially average base-year (min(year)) scenario pollutant raster data to the grid of statpop data of the respective year}
\usage{
prepare_rasterdata_aq_base(data_raster_bfs, data_raster_aq, base_year)
}
\arguments{
\item{data_raster_bfs}{}

\item{data_raster_aq}{}

\item{years}{}
}
\description{
Spatially average base-year (min(year)) scenario pollutant raster data to the grid of statpop data of the respective year
}
