% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_all_raster_data}
\alias{read_all_raster_data}
\title{Reads spatial air quality raster data from Canton Zurich geolion wcs api}
\usage{
read_all_raster_data(ressources, years, boundary)
}
\arguments{
\item{boundary}{}

\item{na_value}{}
}
\description{
Reads spatial air quality raster data from Canton Zurich geolion wcs api
}
