% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_helpers.R
\name{read_bafu_shp}
\alias{read_bafu_shp}
\title{Read *.shp from swisstopo / BAFU}
\usage{
read_bafu_shp(file)
}
\arguments{
\item{file}{}
}
\description{
Read *.shp from swisstopo / BAFU
}
\keyword{internal}
