% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_bfs_life_expectancy_data}
\alias{read_bfs_life_expectancy_data}
\title{Reads Swiss BFS life expectancy data ("Kohortensterbetafeln) from official api}
\usage{
read_bfs_life_expectancy_data(destination_path = "inst/extdata")
}
\arguments{
\item{destination_path}{}
}
\description{
Reads Swiss BFS life expectancy data ("Kohortensterbetafeln) from official api
}
