% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo-admin-cache.R
\name{read_parquet_columns}
\alias{read_parquet_columns}
\title{Read selected columns of a parquet file}
\usage{
read_parquet_columns(path, columns)
}
\arguments{
\item{path}{Local file path.}

\item{columns}{Columns to read.}
}
\value{
A data frame.
}
\description{
Read selected columns of a parquet file
}
\keyword{internal}
