% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_helpers.R
\name{read_statpop_csv}
\alias{read_statpop_csv}
\title{Read *.csv from BFS statpop}
\usage{
read_statpop_csv(file, year, crs = 2056)
}
\arguments{
\item{file}{}

\item{year}{}

\item{crs}{}
}
\description{
Read *.csv from BFS statpop
}
\keyword{internal}
