% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode.R
\name{shortpollutant}
\alias{shortpollutant}
\title{Short pollutant code for a parameter code}
\usage{
shortpollutant(x)
}
\arguments{
\item{x}{Character vector of parameter codes.}
}
\value{
Character vector of pollutant codes.
}
\description{
Collapses the ozone metrics onto the pollutant they describe.
}
\examples{
shortpollutant(c("O3_max_98p_m1", "NO2"))

}
