% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_helpers.R
\name{simplify_aq_rasterdata}
\alias{simplify_aq_rasterdata}
\title{Convert rasterdata into long format tibble}
\usage{
simplify_aq_rasterdata(data)
}
\arguments{
\item{data}{}
}
\description{
Convert rasterdata into long format tibble
}
\keyword{internal}
