% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster-align.R
\name{source_overlap_mask}
\alias{source_overlap_mask}
\title{Mark target cells that overlap the extent of the source by area}
\usage{
source_overlap_mask(x, template, tolerance = 1e-06)
}
\arguments{
\item{x}{Source \code{stars} object.}

\item{template}{Target \code{stars} object.}

\item{tolerance}{Relative tolerance, as a fraction of the cell size.}
}
\value{
A logical matrix the size of the template grid (x, y).
}
\description{
Mark target cells that overlap the extent of the source by area
}
\keyword{internal}
