% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster-cube.R
\name{stack_group}
\alias{stack_group}
\title{Stack one group of rasters sharing a grid}
\usage{
stack_group(rasters, years)
}
\arguments{
\item{rasters}{List of \code{stars} objects.}

\item{years}{Their years.}
}
\value{
A \code{stars} cube with a \code{year} dimension.
}
\description{
Stack one group of rasters sharing a grid
}
\keyword{internal}
