% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{timeseriespars}
\alias{timeseriespars}
\title{Wrapper to supply pollutant-specific list of parameters for timeseries plotting}
\usage{
timeseriespars(parameter)
}
\arguments{
\item{parameter}{}
}
\description{
Wrapper to supply pollutant-specific list of parameters for timeseries plotting
}
\keyword{internal}
