% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster-cube.R
\name{with_dimensions}
\alias{with_dimensions}
\title{Reassemble the attributes of a stars object onto given dimensions}
\usage{
with_dimensions(x, dimensions)
}
\arguments{
\item{x}{A \code{stars} object.}

\item{dimensions}{Dimensions to attach, from \code{\link[stars:st_dimensions]{stars::st_dimensions()}}.}
}
\value{
A \code{stars} object.
}
\description{
Reassemble the attributes of a stars object onto given dimensions
}
\keyword{internal}
