% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate.R
\name{aggregate_emmissions}
\alias{aggregate_emmissions}
\title{Aggregates emission budget data}
\usage{
aggregate_emmissions(data)
}
\arguments{
\item{data}{}
}
\description{
Merges subsectors with small emissions into one common category and sums up emissions accordingly.
}
