% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_helpers.R
\name{aggregate_groups}
\alias{aggregate_groups}
\title{General function do aggregate some statistics from long data}
\usage{
aggregate_groups(
  data,
  y,
  groups,
  nmin = 3,
  perc = list(ymin = 0.05, lower = 0.25, middle = 0.5, upper = 0.75, ymax = 0.95)
)
}
\arguments{
\item{data}{}

\item{y}{}

\item{groups}{}

\item{nmin}{}

\item{perc}{}
}
\description{
General function do aggregate some statistics from long data
}
\keyword{internal}
