% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_helpers.R
\name{download_file}
\alias{download_file}
\title{Get wcs dataset metadata from Canton Zurich geolion api}
\usage{
download_file(download_url, destination_path, file_ext)
}
\arguments{
\item{download_url}{}

\item{destination_path}{}

\item{file_ext}{}

\item{wcs_stack}{}

\item{version}{}
}
\description{
Get wcs dataset metadata from Canton Zurich geolion api
}
\keyword{Download}
\keyword{file}
\keyword{internal}
\keyword{into}
\keyword{tempfile}
