% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_helpers.R
\name{download_zip}
\alias{download_zip}
\title{Download *.zip into tempfile & unzip & delete}
\usage{
download_zip(
  download_url,
  destination_path,
  file_filter = NULL,
  file_ext = ".zip"
)
}
\arguments{
\item{file_ext}{}
}
\description{
Download *.zip into tempfile & unzip & delete
}
\keyword{internal}
