% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_helpers.R
\name{get_bfs_statpop_metadata}
\alias{get_bfs_statpop_metadata}
\title{Wrapper to get statpop dataset download url based on year}
\usage{
get_bfs_statpop_metadata(year)
}
\arguments{
\item{year}{}
}
\description{
Wrapper to get statpop dataset download url based on year
}
\keyword{internal}
