% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_helpers.R
\name{get_geo_admin_metadata}
\alias{get_geo_admin_metadata}
\title{Get dataset asset-metadata from geo.admin api}
\usage{
get_geo_admin_metadata(
  collection,
  filter = ".tif",
  stac_version = "0.9",
  max_check_more = 3
)
}
\arguments{
\item{max_check_more}{}
}
\description{
Get dataset asset-metadata from geo.admin api
}
\keyword{internal}
