% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_helpers.R
\name{get_geo_admin_metadata}
\alias{get_geo_admin_metadata}
\title{read wcs layer from geolion}
\usage{
get_geo_admin_metadata(id, stac_version = "0.9")
}
\arguments{
\item{id}{}

\item{coverage}{}

\item{na_value}{}

\item{cov_stack}{}
}
\description{
read wcs layer from geolion
}
\keyword{Convert}
\keyword{Get}
\keyword{a}
\keyword{api}
\keyword{coverage}
\keyword{data.frame}
\keyword{dataset}
\keyword{from}
\keyword{geo.admin}
\keyword{geolion}
\keyword{internal}
\keyword{list}
\keyword{metadata}
\keyword{to}
\keyword{wcs}
