% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_helpers.R
\name{get_geolion_wfs_metadata}
\alias{get_geolion_wfs_metadata}
\title{Get wfs dataset metadata from Canton Zurich geolion api}
\usage{
get_geolion_wfs_metadata(
  apiurl,
  type = "ms:gem_grenzen",
  version = "2.0.0",
  crs = 2056
)
}
\arguments{
\item{apiurl}{}

\item{type}{}

\item{version}{}

\item{crs}{}
}
\description{
Get wfs dataset metadata from Canton Zurich geolion api
}
\keyword{internal}
