% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_helpers.R
\name{get_opendataswiss_metadata}
\alias{get_opendataswiss_metadata}
\title{Get dataset metadata from opendata.swiss api}
\usage{
get_opendataswiss_metadata(
  apiurl,
  file_filter = ".csv",
  useragent = "Amt für Abfall, Wasser, Energie und Luft, Kanton Zürich"
)
}
\arguments{
\item{apiurl}{}
}
\description{
Get dataset metadata from opendata.swiss api
}
\keyword{internal}
