% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-catalog.R
\name{get_plot}
\alias{get_plot}
\title{Get one plot from a plot catalog}
\usage{
get_plot(catalog, plot, parameter = NULL, year = NULL)
}
\arguments{
\item{catalog}{Plot catalog as built by \code{\link[=plot_catalog]{plot_catalog()}}.}

\item{plot}{Name of the plot.}

\item{parameter, year}{Parameter and year of the plot; \code{NULL} if the plot has none.}
}
\value{
The plot. Stops with an error of class \code{plot_catalog_error} unless exactly one plot matches,
naming the plots available.
}
\description{
Get one plot from a plot catalog
}
\examples{
p <- ggplot2::ggplot() + ggplot2::ggtitle("NO2")
catalog <- plot_catalog(list(NO2 = p), "timeseries", names_to = "parameter")
get_plot(catalog, "timeseries", "NO2")

}
