% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{immission_colorscale}
\alias{immission_colorscale}
\title{Provide official colors for plotting air quality raster data}
\usage{
immission_colorscale(...)
}
\arguments{
\item{...}{}
}
\description{
Provide official colors for plotting air quality raster data
}
\keyword{internal}
