% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{immissionscale}
\alias{immissionscale}
\title{Wrapper to supply pollutant-specific raster data color scales for ggplot2}
\usage{
immissionscale(parameter)
}
\arguments{
\item{parameter}{}
}
\description{
Wrapper to supply pollutant-specific raster data color scales for ggplot2
}
\keyword{internal}
