% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{longmetric}
\alias{longmetric}
\title{Recode short pollutant/parameter string to long metric string}
\usage{
longmetric(x)
}
\arguments{
\item{x}{}
}
\description{
Recode short pollutant/parameter string to long metric string
}
\keyword{internal}
