% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{longparameter}
\alias{longparameter}
\title{Recode short parameter string to long metric string}
\usage{
longparameter(x)
}
\arguments{
\item{x}{}
}
\description{
Recode short parameter string to long metric string
}
\keyword{internal}
