% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{longpollutant}
\alias{longpollutant}
\title{Recode short pollutant/parameter string to long string}
\usage{
longpollutant(x)
}
\arguments{
\item{x}{}
}
\description{
Recode short pollutant/parameter string to long string
}
\keyword{internal}
