% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{pal_emissions}
\alias{pal_emissions}
\title{Provides custom color scales for emissions}
\usage{
pal_emissions(n, name)
}
\arguments{
\item{name}{}
}
\description{
Provides custom color scales for emissions
}
\keyword{internal}
