% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.R
\name{prepare_emmissions}
\alias{prepare_emmissions}
\title{Prepares emission dataset for appropriate aggregation}
\usage{
prepare_emmissions(
  data,
  filter_args = canton == "ZH" & year <= lubridate::year(Sys.Date()) & emission != 0 &
    !(subsector \%in\% c("Weitere Punktquellen OL", "Rheinschifffahrt",
    "Flugverkehr Genf"))
)
}
\arguments{
\item{data}{}

\item{filter_args}{}
}
\description{
Prepares emission dataset for appropriate aggregation
}
