% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_bafu_raster_data}
\alias{read_bafu_raster_data}
\title{Reads raster data from official swisstopo api, used for BAFU data on air pollutants and sensitive ecosystem nitrogen deposition CLE exceedance}
\usage{
read_bafu_raster_data(id, years_filter, boundary, crs = 2056)
}
\arguments{
\item{id}{}

\item{boundary}{}

\item{crs}{}
}
\description{
Reads raster data from official swisstopo api, used for BAFU data on air pollutants and sensitive ecosystem nitrogen deposition CLE exceedance
}
