% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_geolion_wfs}
\alias{read_geolion_wfs}
\title{Reads spatial polygon data from Canton Zurich geolion wfs api}
\usage{
read_geolion_wfs(apiurl, version = "2.0.0", crs = 2056)
}
\arguments{
\item{crs}{}
}
\description{
Reads spatial polygon data from Canton Zurich geolion wfs api
}
