% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_local_csv}
\alias{read_local_csv}
\title{Reads local *.csv file}
\usage{
read_local_csv(
  file,
  delim = ";",
  locale = readr::locale(encoding = "latin1", tz = "Etc/GMT-1"),
  ...
)
}
\arguments{
\item{file}{}

\item{delim}{}

\item{locale}{}

\item{...}{}
}
\description{
Reads local *.csv file
}
