% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_opendataswiss}
\alias{read_opendataswiss}
\title{Reads datasets (*.csv, *.px) from url provided by opendata.swiss}
\usage{
read_opendataswiss(url, source, file_filter = ".csv")
}
\arguments{
\item{url}{}

\item{source}{}

\item{file_filter}{}
}
\description{
Reads datasets (*.csv, *.px) from url provided by opendata.swiss
}
