% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_statpop_raster_data}
\alias{read_statpop_raster_data}
\title{Reads Swiss BFS inhabitant population raster data from official api}
\usage{
read_statpop_raster_data(year, destination_path, boundary, crs = 2056)
}
\arguments{
\item{crs}{}
}
\description{
Reads Swiss BFS inhabitant population raster data from official api
}
