% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{round_off}
\alias{round_off}
\title{Round to nearest whole number (kaufmännisches Runden)}
\usage{
round_off(x, digits = 0)
}
\arguments{
\item{x}{}

\item{digits}{}
}
\description{
Round to nearest whole number (kaufmännisches Runden)
}
\keyword{internal}
