% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{shortpollutant}
\alias{shortpollutant}
\title{Recode parameter string to short pollutant string}
\usage{
shortpollutant(x)
}
\arguments{
\item{x}{}
}
\description{
Recode parameter string to short pollutant string
}
\keyword{internal}
