% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{write_local_csv}
\alias{write_local_csv}
\title{Write local *csv}
\usage{
write_local_csv(data, file, delim = ";", na = "NA", append = FALSE)
}
\arguments{
\item{data}{}

\item{file}{}

\item{delim}{}

\item{na}{}
}
\description{
Write local *csv
}
\keyword{internal}
